Move spinner state from model to tui2 package
Spinner animation state (activeSpinner, SetActiveSpinner, SpinnerFrame,
SpinnerFrameCount, SpinnerTickInterval) is UI rendering logic that was
living in the domain model layer. Moved to internal/tui2/spinnerstate.go.

Removed DisplayForFrame from model.Status — it was only used in model
tests and coupled the domain layer to the spinner package. Moved the
spinner tests to tui2/spinnerstate_test.go.
